Enhance error message extraction and move class to more appropriate namespace.
