+++ src/Providers/Http/Exception/ServerException.php
@@ -6,7 +6,7 @@ namespace WordPress\AiClient\Providers\Http\Exception;
 
 use WordPress\AiClient\Common\Exception\RuntimeException;
 use WordPress\AiClient\Providers\Http\DTO\Response;
-use WordPress\AiClient\Providers\Http\Utilities\ErrorMessageExtractor;
+use WordPress\AiClient\Providers\Http\Util\ErrorMessageExtractor;
 
 /**
  * Exception thrown for 5xx HTTP server errors.
